Assume a multisite DB is ready for the platform
DB stages have been setup in WSUWP Environment that configure our multisite setup ahead of time. This relies on the proper configuration at the wp-config.php level as well.

In the future, it would be nice to find a way to use templating config files and include those with the WSUWP Environment repository as those are pretty specific to the environment and not necessarily the platform that we're building.

Basically this should all work from scratch when we're done.