Add a wp-content directory to be used as the content directory after installation
First. There are some things in the content directory that should not exist until after WordPress has been installed. We should move them over as they are required during provisioning.

Second. content/ is a crappy name for a directory whose original name is wp-content. We should keep it as wp-content so that sites can be moved to other WordPress installations with minimal effort if the need ever arose.