Remove unrequired index.php and xmlrpc.php files
We're running WordPress in a subdirectory in a way that only uses wp-config.php a level up rather than requiring index.php to be there as well. There really isn't a need to have these files as our nginx configuration points to index.php inside the `wordpress` directory and these will never be hit.