Add currently used mu-plugins to docs
All of these are technically in various stages of progress, but implemented, so I'm including them in one primary section.

Once a production release has occured, it may be worth adding an in progress section for mu-plugins. Though - these will likely be developed as optional plugins before making the conversion to must use.