Remove WordPress as a submodule
It seems to make more sense that verified updates to WordPress core
are committed to the platform rather than implemented as a submodule.

While we'd like to avoid running patched WordPress core code in
production, it's always possible that we test something to resolve
an issue that can then be fixed officially.

We'll have to determine the proper system for maintaining this version
of WordPress. Likely most updates will just be rsynced from the core
git repository.