Introduce free-form site domain and path addition
Rather than restricting to subdomains and paths, we should allow
new sites to be created with a full site and path.

http://college.wsu.edu/first/ will create a site with a domain of
college.wsu.edu and a path of first/

We have some basic restrictions to start with and can start
expanding form there.

This is the first in a series of commits to enhance this.