Manage super admins through the $super_admins global
This disables the default core functionality for determining who
is a super admin and provides a custom array of users instead.

By doing this, we grab more granual control over the process. We
also lose the ability to assign super admins on the fly. This is
a good thing, as we can start to create that experience around
the concept of a global admin.