Add scripting to hide upload space options from Network Settings
These are fairly global settings right now that should be changed
carefully. We'll hide them by default and then find the proper
way to display and alter the information at a later time.

This does mean that we'll need to go through existing sites and set
them all to 50MB to match the default.

We'll also need to make sure that settings saved globally will
populate everywhere.