Provide a base for extending a site's allowed filetypes list
Some sites do not need the ability to upload dmg, exe, or zip files

In fact, exe files are a beast of their own that we may not want
to allow in the end. We'll have a trigger available that turns a
site into one with extended file permissions. This will upgrade the
list of allowed filetypes and the amount of space per file allowed
on a per site basis.