Introduce WSUWP_COOKIE_DOMAIN constant for wider auth ability
Build the cookie domain based on the configuration of the WSUWP_COOKIE_DOMAIN
constant. If it is not set, we set cookies to the originally requested domain.
If it is set as auto, we find the root domain of the request and use that for
cookies. If it is set and not explicitly defined as 'auto', we assume that a
cookie domain is being specified, similar to how you would do when using
COOKIE_DOMAIN originally.

See #162