Add custom CSS for admin bar to all pages that use it
This removes the CSS file that we were enqueueing to handle the
CSS changes in the admin. Those changes weren't useful on the front
end. Because the CSS is so short and so unlikely to change, we'll
output it directly in the head rather than worrying about a new
file