Account for individual plugins and themes in build process
Some plugins and themes make sense as part of a larger, distributable
repository that a group can share. Other plugins and themes make
sense as individuals.

Adding an individual directory to our build-plugins and build-themes
setup allows for a more detailed deployment configuration that can
handle the plugins and themes it chooses to.