Add wsuwp:site to the global cache groups
When this was not part of a global cache group, the cache key for
a new site's domain/path was caculated incorrectly and the attempt
to clear a previous 404 request for that same domain/path would
not be successful.

Now, the cached domain/path lookup is cleared properly against the
global group when a new site is added.

Old reproduction behavior:

1. Request http://wsu.edu/this-site-is-invalid/, get 404
2. Setup site http://wsu.edu/this-site-is-invalid/
3. Request http://wsu.edu/this-site-is-invalid/, still get 404

New behavior:

1. Request http://wsu.edu/this-site-is-invalid/, get 404
2. Setup site http://wsu.edu/this-site-is-invalid/
3. Request http://wsu.edu/this-site-is-invalid/, see page!