Clear wsuwp:site domain/path request cache when a site is deleted
Previously we would only clear this cache when updating something
about the site and there was no guarantee it would be deleted when
the site itself was deleted. This could cause confusion if a
replacement site was not accessible after the fact, as a 404 from
the incorrect site may appear instead.

Fixes #271