Test WordPress core #34395, cache output of wp_upload_dir()
On page loads with many images, the use of responsive images may
result in many, many calls to `wp_upload_dir()`. This is an attempt
to properly cache that result for reuse and is a possible patch
for WordPress core in 4.5.

As we have a more complex WP setup than usual, we're testing it in
advance to ensure there are no strange problems.