Add edit_pages and upload_files capabilities to the author role
This allows authors to be assigned to pages through the Editorial
Access Manager plugin.

This was enabled for contributors in #182, which is kind of amusing
to me because contributors have fewer capabilties by default than
authors and we probably should have done it for both from the
beginning. I guess I was looking at solving the submit for review
problem and not focusing on the privileges that EAM provided.