Skip network and global capabilities when counting sites
When determining whether the My Networks or My Sites menu should
display, we count the number of sites that a user is a member of
based on meta keys that track this info. Previously, the check
neglected to account for network and global capabiltiies, so in
most cases users were shown extra menus that were not necessary.