Store and capture a current site menu for proper placement
There are two different scenarios in which we need to capture an
existing site admin menu from the admin bar and display it in its
proper place - after the My Sites menu and after the My Networks
menu.

This abstracts that code a bit so that we can reuse it easily.