Remove forced permalink structure on all sites
It used to be that `/blog/` was forced on every main site of a
network and was difficult to change.

As of https://core.trac.wordpress.org/changeset/34749, this is no
longer the case in WordPress core.

We can remove these filters and restore some capabilities to site
admins that allow for a better URL structure when required.